Fix Code Climate issues
